type merge
