+++ web/apps/photos/src/types/metadata.ts
@@ -1,0 +1,11 @@
+export interface Location {
+    latitude: number;
+    longitude: number;
+}
+
+export interface ParsedExtractedMetadata {
+    location: Location;
+    creationTime: number;
+    width: number;
+    height: number;
+}
